+++ Block/CategoryMapping.php
@@ -46,6 +46,7 @@ use Magento\Store\Model\Store;
 use Magento\Store\Model\StoreManagerInterface;
 use Nosto\Cmp\Logger\LoggerInterface;
 use Nosto\Tagging\Model\Service\Product\Category\DefaultCategoryService as CategoryBuilder;
+use Nosto\Cmp\Helper\Data as NostoHelperData;
 
 class CategoryMapping extends Template
 {
@@ -58,6 +59,9 @@ class CategoryMapping extends Template
     /** @var CategoryBuilder */
     private $categoryBuilder;
 
+    /** @var NostoHelperData */
+    private $nostoHelperData;
+
     /** @var LoggerInterface */
     private $logger;
 
@@ -66,6 +70,7 @@ class CategoryMapping extends Template
      * @param StoreManagerInterface $storeManager
      * @param CollectionFactory $collectionFactory
      * @param CategoryBuilder $categoryBuilder
+     * @param NostoHelperData $nostoHelperData
      * @param Context $context
      * @param LoggerInterface $logger
      */
@@ -73,6 +78,7 @@ class CategoryMapping extends Template
         StoreManagerInterface $storeManager,
         CollectionFactory $collectionFactory,
         CategoryBuilder $categoryBuilder,
+        NostoHelperData $nostoHelperData,
         Context $context,
         LoggerInterface $logger
     ) {
@@ -80,6 +86,7 @@ class CategoryMapping extends Template
         $this->storeManager = $storeManager;
         $this->collectionFactory = $collectionFactory;
         $this->categoryBuilder = $categoryBuilder;
+        $this->nostoHelperData = $nostoHelperData;
         $this->logger = $logger;
     }
 
@@ -118,9 +125,12 @@ class CategoryMapping extends Template
 
             $categories = $this->collectionFactory->create()
                 ->addAttributeToSelect('*')
-                ->addAttributeToFilter('include_in_menu', ['eq' => 1])
                 ->addIsActiveFilter()
                 ->setStore($store);
+
+            if (!$this->nostoHelperData->isAllCategoriesMapEnabled($store)) {
+                $categories->addAttributeToFilter('include_in_menu', ['eq' => 1]);
+            }
 
             /** @var Category $category $item */
             foreach ($categories->getItems() as $category) {