+++ Model/Service/Recommendation/StateAwareCategoryService.php
@@ -172,16 +172,17 @@ class StateAwareCategoryService implements StateAwareCategoryServiceInterface
     ): ?CategoryMerchandisingResult {
         $store = $this->storeManager->getStore();
         $category = $this->getCurrentCategoryString($store);
-        $nostoAccount = $this->accountHelper->findAccount($store);
-        $featureAccess = new FeatureAccess($nostoAccount);
-        if (!$featureAccess->canUseGraphql()) {
-            throw new NostoException('Missing Nosto API_APPS token');
-        }
+        //@phan-suppress-next-line PhanTypeMismatchArgument
         $nostoAccount = $this->accountHelper->findAccount($store);
         if ($nostoAccount === null) {
             throw new NostoException('Account cannot be null');
         }
+        $featureAccess = new FeatureAccess($nostoAccount);
+        if (!$featureAccess->canUseGraphql()) {
+            throw new NostoException('Missing Nosto API_APPS token');
+        }
         // Build filters
+        //@phan-suppress-next-line PhanTypeMismatchArgument
         $this->filterBuilder->init($store);
         $this->filterBuilder->buildFromSelectedFilters(
             $this->state->getActiveFilters()