+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 All notable changes to this project will be documented in this file. This project adheres to Semantic Versioning.
 
+### 3.0.0-rc1
+* Add support for using Elasticsearch as a catalog search engine 
+* Introduce possibility to use all categories (not only the ones in navigation) with Nosto's category merchandising   
+
 ### 2.0.1
 * Change the script type to `application/json` for category mapping
 * Add possibility to debug the category query via Magento's debug logging